// frontend/src/gen/api/src/models/CommentsPhotoResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CommentsListDto } from './CommentsListDto';
import {
    CommentsListDtoFromJSON,
    CommentsListDtoFromJSONTyped,
    CommentsListDtoToJSON,
} from './CommentsListDto';

/**
 * 
 * @export
 * @interface CommentsPhotoResponseDto
 */
export interface CommentsPhotoResponseDto {
    /**
     * 
     * @type {string}
     * @memberof CommentsPhotoResponseDto
     */
    title?: string;
    /**
     * 
     * @type {Array<CommentsListDto>}
     * @memberof CommentsPhotoResponseDto
     */
    comments?: Array<CommentsListDto>;
}

/**
 * Check if a given object implements the CommentsPhotoResponseDto interface.
 */
export function instanceOfCommentsPhotoResponseDto(value: object): boolean {
    return true;
}

export function CommentsPhotoResponseDtoFromJSON(json: any): CommentsPhotoResponseDto {
    return CommentsPhotoResponseDtoFromJSONTyped(json, false);
}

export function CommentsPhotoResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommentsPhotoResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'comments': json['comments'] == null ? undefined : ((json['comments'] as Array<any>).map(CommentsListDtoFromJSON)),
    };
}

export function CommentsPhotoResponseDtoToJSON(value?: CommentsPhotoResponseDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'title': value['title'],
        'comments': value['comments'] == null ? undefined : ((value['comments'] as Array<any>).map(CommentsListDtoToJSON)),
    };
}